refactored lib, unit tests pass
