+++ cmdstanpy/lib.py
@@ -3,13 +3,13 @@ import os.path
 import re
 import tempfile
 
-from typing import List, Dict, Tuple
+from typing import Dict, List
 
 import numpy as np
 import pandas as pd
 
 from cmdstanpy import TMPDIR
-from cmdstanpy.utils import do_command, jsondump, rdump
+from cmdstanpy.utils import jsondump, rdump
 from cmdstanpy.utils import check_csv, cmdstan_path
 
 
@@ -275,17 +275,15 @@ class SamplerArgs(object):
 class RunSet(object):
     """Record of running NUTS sampler on a model."""
 
-    def __init__(self, args: SamplerArgs, chains: int = 2) -> None:
+    def __init__(self, args: SamplerArgs, chains: int = 4) -> None:
         """Initialize object."""
+        self._args = args
         self._chains = chains
-        """number of chains."""
         if chains < 1:
             raise ValueError(
                 'chains must be positive integer value, '
                 'found {i]}'.format(chains)
             )
-        self.args = args
-        """sampler args."""
         self.csv_files = []
         """per-chain sample csv files."""
         if args.output_file is None:
@@ -316,7 +314,13 @@ class RunSet(object):
         ]
         """per-chain sampler command."""
         self._retcodes = [-1 for _ in range(chains)]
-        """per-chain return codes."""
+        self._draws = None
+        self._column_names = None
+        self._num_params = None  # metric dim(s)
+        self._metric_type = None
+        self._metric = None
+        self._stepsize = None
+        self._sample = None
 
     def __repr__(self) -> str:
         repr = 'RunSet(args={}, chains={}'.format(self.args, self._chains)
@@ -327,11 +331,11 @@ class RunSet(object):
 
     @property
     def retcodes(self) -> List[int]:
-        """Get list of retcodes for all chains."""
+        """per-chain return codes."""
         return self._retcodes
 
     def check_retcodes(self) -> bool:
-        """Checks that all chains have retcode 0."""
+        """True when all chains have retcode 0."""
         for i in range(self._chains):
             if self._retcodes[i] != 0:
                 return False
@@ -346,154 +350,19 @@ class RunSet(object):
         self._retcodes[idx] = val
 
     @property
-    def chains(self) -> int:
-        return self._chains
-
-    def check_console_msgs(self) -> bool:
-        """Checks console messages for each chain."""
-        valid = True
-        msg = ""
-        for i in range(self._chains):
-            with open(self.console_files[i], 'r') as fp:
-                contents = fp.read()
-                pat = re.compile(r'^Exception.*$', re.M)
-                errors = re.findall(pat, contents)
-                if len(errors) > 0:
-                    valid = False
-                    msg = '{}chain {}: {}\n'.format(msg, i + 1, errors)
-        if not valid:
-            raise Exception(msg)
-
-    def validate_csv_files(self) -> Dict:
-        """
-        Checks csv output files for all chains.
-        Returns Dict with entries for sampler config and drawset .
-        """
-        dzero = {}
-        for i in range(self._chains):
-            if i == 0:
-                dzero = check_csv(self.csv_files[i])
-            else:
-                d = check_csv(self.csv_files[i])
-                for key in dzero:
-                    if key != 'id' and dzero[key] != d[key]:
-                        raise ValueError(
-                            'csv file header mismatch, '
-                            'file {}, key {} is {}, expected {}'.format(
-                                self.csv_files[i], key, dzero[key], d[key]
-                            )
-                        )
-        dzero['chains'] = self._chains
-        return dzero
-
-
-# TODO:  save sample - mv csv tempfiles to permanent location
-class PosteriorSample(object):
-    """Assembled draws from all chains in a RunSet."""
-
-    def __init__(self, run: Dict = None, csv_files: Tuple[str] = None) -> None:
-        """Initialize object."""
-        self._run = run
-        self._csv_files = csv_files
-        self._sample = None
-        self._metric = None
-        self._stepsize = None
-        if run is None:
-            raise ValueError('missing sampler run info')
-        if csv_files is None:
-            raise ValueError('must specify sampler output csv files')
-        for i in range(len(csv_files)):
-            if not os.path.exists(csv_files[i]):
-                raise ValueError('no such file {}'.format(csv_files[i]))
-        self._chains = run['chains']
-        self._draws = run['draws']
-        self._column_names = run['column_names']
-        self._num_params = run['num_params']
-        self._metric_type = run['metric']
-
-    def __repr__(self) -> str:
-        return 'PosteriorSample(chains={},  draws={}, columns={})'.format(
-            self._chains, self._draws, len(self._column_names)
-        )
-
-    def summary(self) -> pd.DataFrame:
-        """
-        Run cmdstan/bin/stansummary over all output csv files.
-        Echo stansummary stdout/stderr to console.
-        Assemble csv tempfile contents into pandasDataFrame.
-        """
-        names = self.column_names
-        cmd_path = os.path.join(cmdstan_path(), 'bin', 'stansummary')
-        tmp_csv_file = 'stansummary-{}-{}-chains-'.format(
-            self.model, self.chains)
-        fd, tmp_csv_path = tempfile.mkstemp(
-            suffix='.csv', prefix=tmp_csv_file, dir=TMPDIR, text=True
-        )
-        cmd = '{} --csv_file={} {}'.format(
-            cmd_path, tmp_csv_path, ' '.join(self.csv_files)
-        )
-        do_command(cmd.split())  # breaks on all whitespace
-        summary_data = pd.read_csv(
-            tmp_csv_path, delimiter=',', header=0, index_col=0, comment='#'
-        )
-        mask = [
-            x == 'lp__' or not x.endswith('__') for x in summary_data.index
-            ]
-        return summary_data[mask]
-
-    def diagnose(self) -> None:
-        """
-        Run cmdstan/bin/diagnose over all output csv files.
-        Echo diagnose stdout/stderr to console.
-        """
-        cmd_path = os.path.join(cmdstan_path(), 'bin', 'diagnose')
-        csv_files = ' '.join(self.csv_files)
-        cmd = '{} {} '.format(cmd_path, csv_files)
-        result = do_command(cmd=cmd.split())
-        if result is None:
-            print('No problems detected.')
-        else:
-            print(result)
-
-    def extract(self, params: List[str] = None) -> pd.DataFrame:
-        """
-        Returns the assembled sample as a pandas DataFrame consisting of
-        one column per parameter and one row per draw.
-        """
-        pnames_base = [name.split('.')[0] for name in self._column_names]
-        if params is not None:
-            for p in params:
-                if not (p in self._column_names or p in pnames_base):
-                    raise ValueError('unknown parameter: {}'.format(p))
-        if self._sample is None:
-            self.parse_sample()
-        data = self._sample.reshape(
-            (self._draws * self._chains), len(self._column_names), order='A'
-        )
-        df = pd.DataFrame(data=data, columns=self._column_names)
-        if params is None:
-            return df
-        mask = []
-        for p in params:
-            for name in self._column_names:
-                if p == name or p == name.split('.')[0]:
-                    mask.append(name)
-        return df[mask]
-
-    @property
     def model(self) -> str:
         """Stan model name"""
-        return self._run['model']
+        return self._args.model.name
+
+    @property
+    def chains(self) -> int:
+        """Number of sampler chains."""
+        return self._chains
 
     @property
     def draws(self) -> int:
         """Number of draws per chain"""
         return self._draws
-
-    @property
-    def chains(self) -> int:
-        """Number of chains run by sampler"""
-        return self._chains
 
     @property
     def columns(self) -> int:
@@ -513,13 +382,6 @@ class PosteriorSample(object):
         return self._column_names
 
     @property
-    def csv_files(self) -> (str, ...):
-        """
-        Full path name to stan_csv files returned by sampler.
-        """
-        return self._csv_files
-
-    @property
     def metric_type(self) -> str:
         """Metric type, either 'diag_e' or 'dense_e'"""
         return self._metric_type
@@ -528,14 +390,14 @@ class PosteriorSample(object):
     def metric(self) -> np.ndarray:
         """Array of per-chain metric used by sampler."""
         if self._metric is None:
-            self.parse_sample()
+            self.assemble_sample()
         return self._metric
 
     @property
     def stepsize(self) -> np.ndarray:
         """Array or per-chain stepsize."""
         if self._stepsize is None:
-            self.parse_sample()
+            self.assemble_sample()
         return self._stepsize
 
     @property
@@ -547,13 +409,53 @@ class PosteriorSample(object):
         all draws from a chain are contiguous.
         """
         if self._sample is None:
-            self.parse_sample()
+            self.assemble_sample()
         return self._sample
 
-    def parse_sample(self) -> None:
+    def check_console_msgs(self) -> bool:
+        """Checks console messages for each chain."""
+        valid = True
+        msg = ""
+        for i in range(self._chains):
+            with open(self.console_files[i], 'r') as fp:
+                contents = fp.read()
+                pat = re.compile(r'^Exception.*$', re.M)
+                errors = re.findall(pat, contents)
+                if len(errors) > 0:
+                    valid = False
+                    msg = '{}chain {}: {}\n'.format(msg, i + 1, errors)
+        if not valid:
+            raise Exception(msg)
+
+    def validate_csv_files(self) -> None:
+        """
+        Checks that csv output files for all chains are consistent.
+        Populates attributes for drawset, metric.
+        Raises exception when inconsistencies detected.
+        """
+        dzero = {}
+        for i in range(self._chains):
+            if i == 0:
+                dzero = check_csv(self.csv_files[i])
+            else:
+                d = check_csv(self.csv_files[i])
+                for key in dzero:
+                    if key != 'id' and dzero[key] != d[key]:
+                        raise ValueError(
+                            'csv file header mismatch, '
+                            'file {}, key {} is {}, expected {}'.format(
+                                self.csv_files[i], key, dzero[key], d[key]
+                            )
+                        )
+        self._draws = dzero['draws']
+        self._column_names = dzero['column_names']
+        self._num_params = dzero['num_params']
+        self._metric_type = dzero['metric']
+
+    def assemble_sample(self) -> None:
         """
         Allocates and populates the stepsize, metric, and drawset arrays
-        by parsing the validated stan_csv files if they don't already exist.
+        by parsing the validated stan_csv files.
         """
         if not (self._stepsize is None and self._metric is None and
                     self._sample is None):
@@ -571,7 +473,7 @@ class PosteriorSample(object):
             order='F'
             )
         for chain in range(self._chains):
-            with open(self._csv_files[chain], 'r') as fp:
+            with open(self.csv_files[chain], 'r') as fp:
                 # read past initial comments, column header
                 line = fp.readline().strip()
                 while len(line) > 0 and line.startswith('#'):
